Create screens folder
