+++ client/src/actions/completedTrailsActions.js
@@ -1,36 +1,16 @@
-import { FETCH_COMPLETED_TRAILS } from './types';
+import { FETCH_COMPLETED_TRAILS, USER_START_FETCHING } from './types';
 import firebase from '../firebase';
-
-export const getCompletedTrails = (trails) => {
-    return { type: FETCH_COMPLETED_TRAILS, payload: trails}
-}
 
 export const fetchCompletedTrails = () => {
     return (dispatch) => {
-        dispatch(fetchingUserTrailData());
+        dispatch({ type: USER_START_FETCHING });
         if (firebase.auth().currentUser) {
             let userID = firebase.auth().currentUser.uid;
             firebase.database().ref().child('users/' + userID + '/completed')
             .on('value', (snapshot) => {
                     const trails = snapshot.val() || [];
-                    dispatch(getCompletedTrails(trails))
+                    dispatch({ type: FETCH_COMPLETED_TRAILS, payload: trails})
             });
         }
     };
 }
-
-export const deleteCompletedTrail = (trailID) => {
-    return dispatch => {
-        let userID = firebase.auth().currentUser.uid;
-        firebase.database().ref().child('users/' + userID + '/completed/' + trailID).remove()
-    }
-};
-
-export const addCompletedTrail = (trail) => {
-    return dispatch => {
-        let userID = firebase.auth().currentUser.uid;
-        const newTrailRef = firebase.database().ref().child('users/' + userID + '/completed').push()
-        trail.id = newTrailRef.key;
-        newTrailRef.set(trail);
-    }
-};
