+++ client/src/components/screens/ProfileScreen.js
@@ -1,10 +1,9 @@
 import React from 'react';
-import { StyleSheet } from 'react-native';
-import Button from './reusable/Button';
-import LogInRedirect from './LogInRedirect';
+import Button from '../reusable/Button';
+import LogInRedirect from '../LogInRedirect';
 import { connect } from 'react-redux';
-import  { signOutUser } from '../actions';
-import Card from './reusable/Card';
+import  { signOutUser } from '../../actions';
+import Card from '../reusable/Card';
 
 class ProfileScreen extends React.Component {
     signOut() {
